Some things did not change over so this is for porting purposes.

+++ src/main/java/Main.java
@@ -17,7 +17,7 @@ public class Main {
         ChatUtils.println("");
         
         try {
-            int rawPin = Integer.parseInt(pin);
+            long rawPin = Long.parseLong(pin);
             if (rawPin < 0){
                 ChatUtils.println(ParseText.getText( "NegativePin"));
                 Thread.sleep(100);
